Updated Bonds bounded context
Implemented more methods and added Bonds to the Shared bounded context

+++ EBono-API/Bonds/Domain/Models/EExpirationType.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 
 namespace EBono_API.Bonds.Domain.Models
 {
-    public enum EExpirationType : byte
+    public enum EExpirationType : int
     {
         [Description("YR")] 
         Years = 1,
